Changed low warning to be lower and animation case to scale beyond 100 psi
Sometimes the car reads slightly above 100 psi.  In this case it was showing the needle at 0.  This is incorrect so the default animation frame was switched to the 20th frame (needle all the way right).  If the sender is disconnected and a negative number is displayed the user will still be able to tell because the inverted colors on the screen will alert them.

Changed the low warning to be 15.  The car reads about 18 or 19 and with the threshold at 18 the screen flickers between a normal and warning state in an annoying way.  15 psi is the minimum from the FSM for the Subaru wrx sti.